# docs/swagger.yaml
basePath: /
definitions:
...
  models.EscalationRule:
    description: Rule for automatic ticket escalation based on conditions
    properties:
      condition_type:
        description: '"sla_breach", "no_response", "no_update"'
        example: no_response
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        example: Escalate high priority tickets with no response after 30 minutes
        type: string
      escalate_after:
        description: Minutes after creation/last update
        example: 30
        type: integer
      escalate_to:
        description: Escalation Action
        example: 507f1f77bcf86cd799439011
        type: string
      escalate_to_type:
        description: '"user" or "group"'
        example: user
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      is_active:
        description: Status
        example: true
        type: boolean
      name:
        example: High Priority No Response
        type: string
      notify_emails:
        example:
        - manager@example.com
        - support@example.com
        items:
          type: string
        type: array
      priority:
        allOf:
        - $ref: '#/definitions/models.TicketPriority'
        description: Conditions
        example: high
      status:
        allOf:
        - $ref: '#/definitions/models.TicketStatus'
        description: Apply to specific status
        example: open
      updated_at:
        example: "2024-01-13T10:00:00Z"
        type: string
    type: object
...
  models.TicketPriority:
    enum:
    - low
    - medium
    - high
    - urgent
    type: string
    x-enum-varnames:
    - TicketPriorityLow
    - TicketPriorityMedium
    - TicketPriorityHigh
    - TicketPriorityUrgent
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: This is a sample server using Fiber, Uber Fx, and GORM.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Microservice Demo API
  version: "1.0"
paths:
...
  /api/escalation-rules:
    get:
      description: Get all escalation rules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EscalationRule'
            type: array
      summary: List escalation rules
      tags:
      - escalation-rules
    post:
      consumes:
      - application/json
      description: Create a new escalation rule
      parameters:
      - description: Escalation Rule Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.EscalationRule'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Create escalation rule
      tags:
      - escalation-rules
  /api/escalation-rules/{id}:
    delete:
      description: Delete an escalation rule by ID
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Delete escalation rule
      tags:
      - escalation-rules
    get:
      description: Get an escalation rule by ID
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EscalationRule'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get escalation rule
      tags:
      - escalation-rules
    put:
      consumes:
      - application/json
      description: Update an escalation rule by ID
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Data
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Update escalation rule
      tags:
      - escalation-rules
...
  /api/sla-policies:
    get:
      description: Get all SLA policies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SLAPolicy'
            type: array
      summary: List SLA policies
      tags:
      - sla-policies
    post:
      consumes:
      - application/json
      description: Create a new SLA policy
      parameters:
      - description: SLA Policy Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SLAPolicy'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Create SLA policy
      tags:
      - sla-policies
  /api/sla-policies/{id}:
    delete:
      description: Delete an SLA policy by ID
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Delete SLA policy
      tags:
      - sla-policies
    get:
      description: Get an SLA policy by ID
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SLAPolicy'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get SLA policy
      tags:
      - sla-policies
    put:
      consumes:
      - application/json
      description: Update an SLA policy by ID
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Data
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Update SLA policy
      tags:
      - sla-policies
  /api/tickets:
    get:
      consumes:
      - application/json
      description: Get tickets with filtering and pagination
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Records per page (default 10)
        in: query
        name: limit
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by priority
        in: query
        name: priority
        type: string
      - description: Filter by channel
        in: query
        name: channel
        type: string
      - description: Search by ticket number, subject, or customer
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Ticket'
            type: array
        "400":
          description: Invalid input
          schema:
            type: string
      summary: List tickets
      tags:
      - tickets
    post:
      consumes:
      - application/json
      description: Create a new support ticket
      parameters:
      - description: Ticket Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Ticket'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Create a new ticket
      tags:
      - tickets
  /api/tickets/{id}:
    delete:
      description: Delete a ticket by ID
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Delete a ticket
      tags:
      - tickets
    get:
      description: Get a ticket by ID
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ticket'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get a ticket
      tags:
      - tickets
    put:
      consumes:
      - application/json
      description: Update a ticket by ID
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Data
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Update a ticket
      tags:
      - tickets
  /api/tickets/{id}/assign:
    patch:
      consumes:
      - application/json
      description: Assign a ticket to a user
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      - description: Assignment Data
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Assign ticket
      tags:
      - tickets
  /api/tickets/{id}/comments:
    get:
      description: Get all comments for a ticket
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TicketComment'
            type: array
        "400":
          description: Invalid input
          schema:
            type: string
      summary: List ticket comments
      tags:
      - tickets
    post:
      consumes:
      - application/json
      description: Add a comment or note to a ticket
      parameters:
      - description: Ticket ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.TicketComment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
      summary: Add comment to ticket
      tags:
      - tickets
...
swagger: "2.0"